Refactored after utility -> utility2
