+++ src/execution/ExecutionPlan.cpp
@@ -8,8 +8,8 @@
 
 #include "scheduler/SchedulerTypedefs.hpp"
 #include "scheduler/SchedulerInterface.hpp"
-#include "utility/EventProfiler.hpp"
-#include "utility/SyncStream.hpp"
+#include "utility2/EventProfiler.hpp"
+#include "utility2/SyncStream.hpp"
 #include "execution/ExecutionPlan.hpp"
 
 #include "gflags/gflags.h"
