+++ src/scheduler/CMakeLists.txt
@@ -1,72 +1,72 @@
 # Declare micro-libs:
-add_library(project_scheduler_Node ../empty_src.cpp Node.hpp)
-add_library(project_scheduler_Scheduler Scheduler.cpp Scheduler.hpp)
-add_library(project_scheduler_SchedulerFlags SchedulerFlags.cpp SchedulerFlags.hpp)
-add_library(project_scheduler_SchedulerInterface ../empty_src.cpp SchedulerInterface.hpp)
-add_library(project_scheduler_SchedulerMessage ../empty_src.cpp SchedulerMessage.hpp)
-add_library(project_scheduler_SchedulerTypedefs ../empty_src.cpp SchedulerTypedefs.hpp)
-add_library(project_scheduler_Task ../empty_src.cpp Task.hpp)
-add_library(project_scheduler_TaskDescription ../empty_src.cpp TaskDescription.hpp)
-add_library(project_scheduler_TaskName ../empty_src.cpp TaskName.hpp)
-add_library(project_scheduler_TaskStatistics TaskStatistics.cpp TaskStatistics.hpp)
-add_library(project_scheduler_Worker Worker.cpp Worker.hpp)
-add_library(project_scheduler_WorkerMessage ../empty_src.cpp WorkerMessage.hpp)
+add_library(hustle_src_scheduler_Node ../empty_src.cpp Node.hpp)
+add_library(hustle_src_scheduler_Scheduler Scheduler.cpp Scheduler.hpp)
+add_library(hustle_src_scheduler_SchedulerFlags SchedulerFlags.cpp SchedulerFlags.hpp)
+add_library(hustle_src_scheduler_SchedulerInterface ../empty_src.cpp SchedulerInterface.hpp)
+add_library(hustle_src_scheduler_SchedulerMessage ../empty_src.cpp SchedulerMessage.hpp)
+add_library(hustle_src_scheduler_SchedulerTypedefs ../empty_src.cpp SchedulerTypedefs.hpp)
+add_library(hustle_src_scheduler_Task ../empty_src.cpp Task.hpp)
+add_library(hustle_src_scheduler_TaskDescription ../empty_src.cpp TaskDescription.hpp)
+add_library(hustle_src_scheduler_TaskName ../empty_src.cpp TaskName.hpp)
+add_library(hustle_src_scheduler_TaskStatistics TaskStatistics.cpp TaskStatistics.hpp)
+add_library(hustle_src_scheduler_Worker Worker.cpp Worker.hpp)
+add_library(hustle_src_scheduler_WorkerMessage ../empty_src.cpp WorkerMessage.hpp)
 
 # Link dependencies:
-target_link_libraries(project_scheduler_Node
+target_link_libraries(hustle_src_scheduler_Node
                       glog
-                      project_scheduler_SchedulerTypedefs
-                      project_utility_ContainerUtil
-                      project_utility_Macros)
-target_link_libraries(project_scheduler_Scheduler
+                      hustle_src_scheduler_SchedulerTypedefs
+                      hustle_src_utility_ContainerUtil
+                      hustle_src_utility_Macros)
+target_link_libraries(hustle_src_scheduler_Scheduler
                       glog
-                      project_scheduler_Node
-                      project_scheduler_SchedulerFlags
-                      project_scheduler_SchedulerInterface
-                      project_scheduler_SchedulerMessage
-                      project_scheduler_SchedulerTypedefs
-                      project_scheduler_Task
-                      project_scheduler_Worker
-                      project_scheduler_WorkerMessage
-                      project_utility_ContainerUtil
-                      project_utility_Macros
-                      project_utility_ThreadSafeQueue)
-target_link_libraries(project_scheduler_SchedulerFlags
+                      hustle_src_scheduler_Node
+                      hustle_src_scheduler_SchedulerFlags
+                      hustle_src_scheduler_SchedulerInterface
+                      hustle_src_scheduler_SchedulerMessage
+                      hustle_src_scheduler_SchedulerTypedefs
+                      hustle_src_scheduler_Task
+                      hustle_src_scheduler_Worker
+                      hustle_src_scheduler_WorkerMessage
+                      hustle_src_utility_ContainerUtil
+                      hustle_src_utility_Macros
+                      hustle_src_utility_ThreadSafeQueue)
+target_link_libraries(hustle_src_scheduler_SchedulerFlags
                       ${GFLAGS_LIB_NAME})
-target_link_libraries(project_scheduler_SchedulerInterface
-                      project_scheduler_SchedulerTypedefs
-                      project_utility_Macros)
-target_link_libraries(project_scheduler_SchedulerMessage
-                      project_utility_Macros)
-target_link_libraries(project_scheduler_Task
+target_link_libraries(hustle_src_scheduler_SchedulerInterface
+                      hustle_src_scheduler_SchedulerTypedefs
+                      hustle_src_utility_Macros)
+target_link_libraries(hustle_src_scheduler_SchedulerMessage
+                      hustle_src_utility_Macros)
+target_link_libraries(hustle_src_scheduler_Task
                       glog
-                      project_scheduler_SchedulerInterface
-                      project_scheduler_SchedulerTypedefs
-                      project_scheduler_TaskDescription
-                      project_utility_Macros
-                      project_utility_meta_FunctionTraits)
-target_link_libraries(project_scheduler_TaskDescription
+                      hustle_src_scheduler_SchedulerInterface
+                      hustle_src_scheduler_SchedulerTypedefs
+                      hustle_src_scheduler_TaskDescription
+                      hustle_src_utility_Macros
+                      hustle_src_utility_meta_FunctionTraits)
+target_link_libraries(hustle_src_scheduler_TaskDescription
                       glog
-                      project_scheduler_TaskName
-                      project_utility_Macros)
-target_link_libraries(project_scheduler_TaskName
-                      project_utility_Macros)
-target_link_libraries(project_scheduler_TaskStatistics
-                      project_scheduler_Scheduler
-                      project_scheduler_TaskDescription
-                      project_utility_Macros)
-target_link_libraries(project_scheduler_Worker
+                      hustle_src_scheduler_TaskName
+                      hustle_src_utility_Macros)
+target_link_libraries(hustle_src_scheduler_TaskName
+                      hustle_src_utility_Macros)
+target_link_libraries(hustle_src_scheduler_TaskStatistics
+                      hustle_src_scheduler_Scheduler
+                      hustle_src_scheduler_TaskDescription
+                      hustle_src_utility_Macros)
+target_link_libraries(hustle_src_scheduler_Worker
                       ${GFLAGS_LIB_NAME}
                       glog
-                      project_scheduler_SchedulerInterface
-                      project_scheduler_SchedulerMessage
-                      project_scheduler_Task
-                      project_scheduler_WorkerMessage
-                      project_threading_ThreadUtil
-                      project_utility_Macros
-                      project_utility_ThreadSafeQueue)
-target_link_libraries(project_scheduler_WorkerMessage
-                      project_utility_Macros)
+                      hustle_src_scheduler_SchedulerInterface
+                      hustle_src_scheduler_SchedulerMessage
+                      hustle_src_scheduler_Task
+                      hustle_src_scheduler_WorkerMessage
+                      hustle_src_threading_ThreadUtil
+                      hustle_src_utility_Macros
+                      hustle_src_utility_ThreadSafeQueue)
+target_link_libraries(hustle_src_scheduler_WorkerMessage
+                      hustle_src_utility_Macros)
 
 add_executable(hustle_scheduler_test "tests/SchedulerTest.cpp")
 target_link_libraries(hustle_scheduler_test
@@ -74,7 +74,7 @@ target_link_libraries(hustle_scheduler_test
         gtest_main
         gmock
         ${ARROW_SHARED_LIB}
-        project_scheduler_Scheduler
-        operators
+        hustle_src_scheduler_Scheduler
+        hustle_src_operators
         )
 add_test(HustleSchedulerTest hustle_scheduler_test)
