+++ src/scheduler/Scheduler.cpp
@@ -15,7 +15,7 @@
 #include "../scheduler/SchedulerTypedefs.hpp"
 #include "../scheduler/Task.hpp"
 #include "../scheduler/Worker.hpp"
-#include "../utility/ContainerUtil.hpp"
+#include "../utility2/ContainerUtil.hpp"
 
 namespace hustle {
