+++ src/scheduler/SchedulerInterface.hpp
@@ -4,7 +4,7 @@
 #include <memory>
 
 #include "../scheduler/SchedulerTypedefs.hpp"
-#include "../utility/Macros.hpp"
+#include "../utility2/Macros.hpp"
 
 namespace hustle {
