+++ src/scheduler/WorkerMessage.hpp
@@ -3,7 +3,7 @@
 
 #include <memory>
 
-#include "../utility/Macros.hpp"
+#include "../utility2/Macros.hpp"
 
 namespace hustle {
