+++ src/threading/SpinMutex.hpp
@@ -4,7 +4,7 @@
 #include <atomic>
 
 #include "threading/Mutex.hpp"
-#include "utility/Macros.hpp"
+#include "utility2/Macros.hpp"
 
 namespace hustle {
