+++ src/threading/SynchronizationLock.hpp
@@ -6,7 +6,7 @@
 #include <memory>
 #include <mutex>
 
-#include "utility/Macros.hpp"
+#include "utility2/Macros.hpp"
 
 class SynchronizationLock {
  public:
